% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_import_from_statements.R
\name{add_import_from_statements}
\alias{add_import_from_statements}
\title{Add @importFrom Statements Based on Script Dependencies}
\usage{
add_import_from_statements(script)
}
\arguments{
\item{script}{A character string specifying the path to the R script file.}
}
\value{
This function generates and prints `@importFrom` statements for each
package detected in the script.
}
\description{
This function analyzes a given script to detect functions used in the
`package::function` notation. It then generates `@importFrom` statements
for each package detected, capturing the required dependencies for the script.
}
\examples{
ex_script <- "mtcars \%>\%
   dplyr::mutate(
     var_1 = DescTools::Quantile(disp, probs = .6),
     var_2 = str_glue('{disp} {hp}')
   ) \%>\%
   dplyr::arrange(disp)"
temp_script <- tempfile(fileext = ".R")
writeLines(ex_script, temp_script)
add_import_from_statements(temp_script)


}
