% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_me.R
\name{alert_me}
\alias{alert_me}
\title{Display an alert pop-up screen}
\usage{
alert_me()
}
\description{
This function creates a pop-up screen that alerts the user that a process
has been completed. By default, it generates a big red pop-up screen with
the message "ALERT: SCRIPT HAS COMPLETED".
}
\details{
The alert pop-up screen is created using the ggplot2 package to generate
a simple plot with a label containing the alert message.
}
\examples{
\dontrun{
  alert_me()
}

}
\seealso{
\code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{labs}},
\code{\link[ggplot2]{theme}}, \code{\link[ggplot2]{element_blank}},
\code{\link[ggplot2]{element_rect}}
}
