% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timeseries.R
\name{auto_make_diff_col}
\alias{auto_make_diff_col}
\title{Make multiple differenced columns.}
\usage{
auto_make_diff_col(col, lags = c(1, 7), percent = F)
}
\arguments{
\item{col}{a column to make differences of}

\item{lags}{numeric list}

\item{percent}{boolean to determine if difference should be percent difference}
}
\value{
dataframe with differenced column(s)
}
\description{
Makes differenced column(s) from a dataframe column defined by the user.
}
\examples{
data.frame(index = 1:10,
          value1 = rnorm(10), value2 = rnorm(10)) \%>\%
 mutate(auto_make_diff_col(col = value1, percent = T))
}
