% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_make_lag_col.R
\name{auto_make_lag_col}
\alias{auto_make_lag_col}
\title{Make multiple lagged columns.}
\usage{
auto_make_lag_col(col, lags = c(1, 7))
}
\arguments{
\item{col}{a column to make lags of}

\item{lags}{numeric list}
}
\value{
dataframe with lagged column(s)
}
\description{
Makes lagged column(s) from a dataframe column defined by the user.
}
\examples{

\dontrun{
data.frame(index = 1:10,
value1 = rnorm(10), value2 = rnorm(10)) \%>\%
mutate(auto_make_lag_col(col = value1, c(1, 3, 5)))
}
}
