% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timeseries.R
\name{auto_make_rMean_col}
\alias{auto_make_rMean_col}
\title{Make multiple rolling mean functions.}
\usage{
auto_make_rMean_col(col, width = c(7))
}
\arguments{
\item{col}{a column to make rolling means of}

\item{width}{integer input indicating size of sliding window used to caluclate mean}
}
\value{
dataframe with rolling mean column(s)S
}
\description{
Makes rolling mean column(s) from a dataframe column defined by the user using window sizes also defined by the user.
}
\examples{
data.frame(index = 1:10,
          value1 = rnorm(10), value2 = rnorm(10)) \%>\%
 mutate(auto_make_rMean_col(col = value1, width = c(3, 7)))
}
