% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_file_manipulation.R
\name{check_dir_path}
\alias{check_dir_path}
\title{Check and create a directory path}
\usage{
check_dir_path(dir_path)
}
\arguments{
\item{dir_path}{A string specifying the path of the directory to check and create if necessary.}
}
\value{
None
}
\description{
This function checks if the directory path exists and creates it if it doesn't.
}
\examples{
check_dir_path("data/raw")
check_dir_path("data/processed")
}
