% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPkg_map_functions_to_scripts.R
\name{checkPkg_map_functions_and_scripts}
\alias{checkPkg_map_functions_and_scripts}
\title{Map Functions and Scripts in an R Package}
\usage{
checkPkg_map_functions_and_scripts(function_directory, code_directory = NULL)
}
\arguments{
\item{function_directory}{Character. Path to the directory containing R scripts with function definitions (e.g., `R` directory in an R package).}

\item{code_directory}{Character. Path to the directory containing other R scripts (e.g., `inst/scripts` directory). Default is \code{NULL}.}
}
\value{
A list with two elements:
  \describe{
    \item{\code{vis_plot}}{An interactive network plot showing the relationships between functions and scripts.}
    \item{\code{df_to_from}}{A data frame mapping function-to-function and function-to-script relationships.}
  }
}
\description{
This function analyzes R scripts within two directories: a function directory (e.g., `R` folder for an R package) and an optional code directory (e.g., `inst/scripts` folder). It maps the relationships between functions and scripts, visualizing the connections in a network diagram and providing a data frame of function calls.
}
\details{
The function performs the following steps:
1. Scans the `function_directory` to identify R scripts and extracts all defined functions.
2. Optionally scans the `code_directory` for additional scripts.
3. Maps function calls within and across scripts, creating a data frame of relationships.
4. Visualizes the relationships in an interactive network plot using \pkg{visNetwork}.

The function can be used to understand code dependencies, identify potential redundancies, and visualize how functions and scripts interact within a project.
}
\note{
This function assumes that R scripts in the `function_directory` contain properly defined functions, and that scripts in the `code_directory` may invoke these functions.
}
\examples{
\dontrun{
checkPkg_map_functions_and_scripts(function_directory = "R", code_directory = "inst/scripts")
}
}
