% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_percent_zscore.R
\name{count_percent_zscore}
\alias{count_percent_zscore}
\title{Group by count, sum, percent and zscore.}
\usage{
count_percent_zscore(
  data,
  grp_c = ...,
  grp_p = ...,
  grp_z = ...,
  col,
  prefix = NULL,
  rnd = NULL,
  cntr_scl = FALSE
)
}
\arguments{
\item{data}{a dataframe}

\item{grp_c}{a vector of columns to group counting operation with - do not quote columns}

\item{grp_p}{a vector of columns to group percent calculation operation with - do not quote columns}

\item{grp_z}{a vector of columns to group zscore calculation operation with - do not quote columns}

\item{col}{a column to count or sum - do not quote column}

\item{prefix}{a string used to prefix calculated columns with - leave empty if you do not want a prefix}

\item{rnd}{integer indicating how many digits you want calculated columns to be rounded to - leave empty if you do not want rounding}

\item{cntr_scl}{(`TRUE`/`FALSE`) boolean to indicate if zscore should be calculated - default is `FALSE`}
}
\value{
a dataframe
}
\description{
This function can be used to perform a number of mathematical operation on a dataframe.
Count or sum numerical values by group, get percent breakdown of counts or sums by group, and get optional zscore of values.
}
\examples{

\dontrun{
temp_data = data.frame(group = c(rep("A", 4), rep("B", 4))
                      ,order = c(1:4, 1:4)) \%>\%
 mutate(value = 2*order+rnorm(8, 5)
        ,count = 1)

count_percent_zscore(temp_data, grp_c = c(group), grp_p = c(), col = count)

count_percent_zscore(temp_data, grp_c = c(order), grp_p = c(), col = value)
}
}
