% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrct0.R
\name{crrct0}
\alias{crrct0}
\title{Rescale a vector with first entry.}
\usage{
crrct0(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Use this function to rescale a vector using the first entry
}
\examples{

\dontrun{
data.frame(group = c(rep("A", 4), rep("B", 4))
          ,order = c(1:4, 1:4)) \%>\%
 mutate(value = 2*order+rnorm(8, 5)) \%>\%
 group_by(group) \%>\%
 mutate(value_corrected = crrct0(value))
 }
}
