% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_manip.R
\name{dgt1}
\alias{dgt1}
\title{Round to 1 digits}
\usage{
dgt1(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Use this function to round a vector.
}
\examples{
dgt1(rnorm(10))
}
