% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_manip.R
\name{dgt2}
\alias{dgt2}
\title{Round to 2 digits}
\usage{
dgt2(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Use this function to round a vector.
}
\examples{
dgt2(rnorm(10))
}
