% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_manip.R
\name{dgt3}
\alias{dgt3}
\title{Round to 3 digits}
\usage{
dgt3(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Use this function to round a vector.
}
\examples{
dgt3(rnorm(10))
}
