% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{flatten_named_list}
\alias{flatten_named_list}
\title{Flatten a named list}
\usage{
flatten_named_list(lst, parent_name = "")
}
\arguments{
\item{lst}{A named list to be flattened into a data frame.}

\item{parent_name}{A character string indicating the name of the parent list
element. This argument is optional and should only be provided if the input
list is a child element of a larger parent list. By default, the parent name
is set to an empty string.}
}
\value{
A data frame with two columns: the name of each list element (including
parent element names), and its corresponding value.
}
\description{
This function takes a named list as input and returns a data frame with two
columns: the name of each list element (including parent element names), and
its corresponding value. If the input is not a list, the function returns a
data frame with a single row containing the value and its parent name (if
provided).
}
\examples{
test_list <- list(a = 1, b = list(c = 2, d = 3))
flatten_named_list(test_list)

}
\keyword{data}
\keyword{frame,}
\keyword{list,}
\keyword{named}
\keyword{recursion}
