% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_logging.R
\name{glue_collapse}
\alias{glue_collapse}
\title{Create a multi-line, printable console message from an object.}
\usage{
glue_collapse(msg)
}
\arguments{
\item{msg}{a string using str_glue syntax to print a df like object}
}
\value{
a string congaing "\n" to force string to print on multiple lines
}
\description{
Function takes glue input and properly collpases it using "\n" to make it multi-line string.
 Using message without properly collapsing will create one long, flat message string.
 Nice when printing small tables.
}
\examples{

temp = data.frame(
highway = c("90 East", "405 Northbound", "101")
,count = c(100, 205, 92)
) \%>\%
 mutate(percent = round(count/sum(count), 2))

str_glue("{temp$count} ({temp$percent}\%) - {temp$highway}") \%>\%
 message()

glue_collapse("{temp$count} ({temp$percent}\%) - {temp$highway}") \%>\%
 message()
}
