% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_manip.R
\name{group_wtd_quantiles}
\alias{group_wtd_quantiles}
\title{Calculate quantiles for groups of pre-aggregated metrics.}
\usage{
group_wtd_quantiles(
  df,
  value,
  weight = "count",
  quantiles = c(0, 0.25, 0.5, 0.75, 1)
)
}
\arguments{
\item{df}{a dataframe of data}

\item{value}{string of column that qunatiles will be calculated for}

\item{weight}{string of column that will be used to calculate quantiles with}

\item{quantiles}{vector of quantiles that should be returned - default is `c(0, .25, .5, .75, 1)` quantiles}
}
\value{
a dataframe with qunatiles - in wide format
}
\description{
Use this function to calculate qunatiles for data that has already been aggregated and collapsed into counts. Avoids the need to expand data then compute quantiles.
}
\examples{

df = data.frame(group = c(rep(1, 4)
                          ,rep(2, 3))
                ,days = c(round(runif(4, 0, 20), 0)
                          ,round(runif(3, 0, 20), 0))
                ,count = c(round(runif(4, 0, 5), 0)
                           ,round(runif(3, 0, 5), 0)))

df \%>\%
  group_by(group) \%>\%
  nest() \%>\%
  mutate(qauntiles = map(data, ~group_wtd_quantiles(.x, value = "days", weight = "count"))) \%>\%
  unnest(cols = c(qauntiles)) \%>\%
  select(!data)

#alternative calucaltion by expansion
rep(df$days[1:4], df$count[1:4]) \%>\%  summary()
}
