% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitting.R
\name{knit_dashboard}
\alias{knit_dashboard}
\title{Automate knitting of RMD FlexDashboards to HTML.}
\usage{
knit_dashboard(
  dashboard_name = "dashboard",
  dashboard_folder = "dashboard",
  doc_name = dashboard_name,
  overwrite = F,
  write_to_folder = "public"
)
}
\arguments{
\item{dashboard_name}{string of RMD to knit - do not include `.RMD` extension.}

\item{dashboard_folder}{string of folder name where RMD is located - default is `analysis`, do not include trailing `(/)`.}

\item{doc_name}{string of knitted RMD name - default is `markdown_name`.}

\item{overwrite}{boolean (`TRUE`/`FALSE`) indicating if file should be rewritten or if date suffix should be added to file name - default is `FALSE` which adds date suffix.}

\item{write_to_folder}{string of folder name where RMD is written to - default is `public`.}
}
\value{
a knitted HTML document
}
\description{
Use this function to knit RMD FlexDashboards to HTMLs and write them to new location and with a different name.
}
\examples{
#none
}
