% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitting.R
\name{knit_markdown}
\alias{knit_markdown}
\title{Automate knitting of RMD to HTML.}
\usage{
knit_markdown(
  markdown_name,
  markdown_folder = "analysis",
  doc_name = markdown_name,
  overwrite = F,
  clean = TRUE,
  write_to_folder = "public"
)
}
\arguments{
\item{markdown_name}{string of RMD to knit - do not include `.RMD` extension.}

\item{markdown_folder}{string of folder name where RMD is located - default is `analysis`.}

\item{doc_name}{string of knitted RMD name - default is `markdown_name`.}

\item{overwrite}{boolean (`TRUE`/`FALSE`) indicating if file should be rewritten or if date suffix should be added to file name - default is `FALSE` which adds date suffix.}

\item{clean}{boolean (`TRUE`/`FALSE`) - keep as default `TRUE`}

\item{write_to_folder}{string of folder name where RMD is written to - default is `public`.}
}
\value{
a knitted HTML document
}
\description{
Use this function to knit RMDs to HTMLs and write them to new location and with a different name.
}
\examples{
#none
}
