% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_leaflet.R
\name{leaflet_default_tiles}
\alias{leaflet_default_tiles}
\title{Default tiles for a leaflet map.}
\usage{
leaflet_default_tiles(object)
}
\arguments{
\item{object}{a blank leaflet object}
}
\value{
a leaflet object
}
\description{
Default tiles for a leaflet map.
}
\examples{
leaflet() \%>\%
  leaflet_default_tiles()
}
