% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_leaflet.R
\name{leaflet_default_tiles_index}
\alias{leaflet_default_tiles_index}
\title{Commonly used Leaflet tiles.}
\usage{
leaflet_default_tiles_index()
}
\value{
Creates leaflet tiles for leaflet map.
}
\description{
Commonly used Leaflet tiles.
}
\examples{
#none
}
