% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_and_fatal.R
\name{log_and_fatal}
\alias{log_and_fatal}
\title{Log and print to console an info message.}
\usage{
log_and_fatal(message, logger)
}
\arguments{
\item{message}{string message to print.}

\item{logger}{a logger object made using log4r package.}
}
\value{
message printed to console and log file with logged message
}
\description{
Log and print to console an info message.
}
\examples{

#none
}
