% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_logging.R
\name{log_and_warn}
\alias{log_and_warn}
\title{Log and print to console a warn message.}
\usage{
log_and_warn(message, logger)
}
\arguments{
\item{message}{string message to print.}

\item{logger}{a logger object made using log4r package.}
}
\value{
message printed to console and log file with logged message
}
\description{
Log and print to console a warn message.
}
\examples{

#none
}
