% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diff.R
\name{make_diff}
\alias{make_diff}
\title{make_diff}
\usage{
make_diff(col, lags, percent = F)
}
\arguments{
\item{col}{a column to make lags}

\item{lags}{numeric list}

\item{percent}{boolean to determine if difference should be percent difference}
}
\value{
dataframe with differenced column
}
\description{
Makes difference column from a dataframe column defined by the user using lags as defined by the user. Should not be used directly by user - it is better to use via auto_make_diff_col() function.
}
\examples{

\dontrun{
data.frame(index = 1:10,
          value1 = rnorm(10), value2 = rnorm(10)) \%>\%
 mutate(value1_differenced = make_diff(col = index, c(1), percent = F)) \%>\%
 mutate(value1_differenced_percent = make_diff(col = index, c(1), percent = T))
 }
}
