% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timeseries.R
\name{make_diff_col}
\alias{make_diff_col}
\title{make_diff_col}
\usage{
make_diff_col(lags = c(1, 7), percent = F)
}
\arguments{
\item{lags}{numeric list}

\item{percent}{boolean to determine if difference should be percent difference}
}
\value{
pre-filled function
}
\description{
Pre-filled difference function. This function should not be directly used, it is used inside of auto_make_diff_col() .
}
\examples{
#none
}
