% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_file_manipulation.R
\name{make_dir_prfx_date}
\alias{make_dir_prfx_date}
\title{Create a directory with a prefix and date}
\usage{
make_dir_prfx_date(directory_location, prefix)
}
\arguments{
\item{directory_location}{A string specifying the location of the directory to create.}

\item{prefix}{A string specifying the prefix to use for the directory name.}
}
\value{
A string specifying the path of the created directory.
}
\description{
This function creates a directory with a specified prefix and the current date.
}
\examples{
make_dir_prfx_date("data/processed", "my_data_")
make_dir_prfx_date("data/raw", "my_data_")
}
