% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lag_col.R
\name{make_lag_col}
\alias{make_lag_col}
\title{make_lag_col}
\usage{
make_lag_col(lags = c(1, 7))
}
\arguments{
\item{lags}{numeric list}
}
\value{
pre-filled function
}
\description{
Pre-filled lag function. This function should not be directly used, it is used inside of auto_make_lag_col function.
}
\examples{
#none
}
