% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_file_manipulation.R
\name{make_new_rmd}
\alias{make_new_rmd}
\title{Make new Rmd script from template.}
\usage{
make_new_rmd(folder = "analysis", file_name = NULL)
}
\arguments{
\item{folder}{a string name (defaults to `analysis` folder)}

\item{file_name}{a string name (include .Rmd extension)}
}
\value{
an Rmd in the location and with the name chosen by the user
}
\description{
This function copies an Rmd script template from the `gauntlet` package and writes it to a new location as defined by the user.
Write-to folder locations must exist before running this code.
}
\examples{
#none
}
