% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_file_manipulation.R
\name{make_new_script}
\alias{make_new_script}
\title{Make new R script from template.}
\usage{
make_new_script(folder = "code", file_name = NULL)
}
\arguments{
\item{folder}{a string name}

\item{file_name}{a string name (include .r extentsion)}
}
\value{
a script in the location and with the name choosen by the user
}
\description{
This function copies an R script template from the `gauntlet` package and writes it to a new location as defined by the user.
Write-to folder locations must exist before running this code.
}
\examples{
#none
}
