% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_timeseries.R
\name{make_rMean_col}
\alias{make_rMean_col}
\title{make_rMean_col}
\usage{
make_rMean_col(width, equal)
}
\arguments{
\item{width}{integer input indicating size of sliding window used to caluclate mean}

\item{equal}{currently not used}
}
\value{
pre-filled function
}
\description{
Pre-filled rolling mean function. This function should not be directly used, it is used inside of auto_make_rMean_col() .
}
\examples{
#none
}
