% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_space.R
\name{make_space}
\alias{make_space}
\title{Make a string using a character repeated some amount of times.}
\usage{
make_space(with = "+", n = 50, c = "")
}
\arguments{
\item{with}{string character to repeat - default is a plus sign ('+')}

\item{n}{integer indicating number of times string character should be repeated - default 50}

\item{c}{string character collapse with input - default is an empty character ('')}
}
\value{
string of collapsed, repeated character
}
\description{
Make a string using a character repeated some amount of times.
}
\details{
This function is deprecated please use `gauntlet::strg_make_space_2` instead.
}
\examples{

make_space(with = "//", c = "\\\\")
}
