% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_leaflet.R
\name{popup_tbl_pretty}
\alias{popup_tbl_pretty}
\title{Create a popup table with formatted content}
\usage{
popup_tbl_pretty(data)
}
\arguments{
\item{data}{A spatial object.}
}
\value{
A leaflet popup table.
}
\description{
This function takes a spatial object and creates a popup table with formatted content.
}
