% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_manip.R
\name{pretty_num}
\alias{pretty_num}
\title{Make Numeric Values Pretty}
\usage{
pretty_num(vector, rnd = 0)
}
\arguments{
\item{vector}{A numeric vector.}

\item{rnd}{Number of decimal places to round to (default is 0).}
}
\value{
A character vector with formatted numeric values.
}
\description{
This function takes a numeric vector and formats it as character numbers with a magnitude suffix.
}
\examples{
pretty_num(c(1000, 1500000, 3.14159, -500000))

}
\author{
Mike Gaunt
}
