% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_file_manipulation.R
\name{print_named_list}
\alias{print_named_list}
\title{Convert named list to string}
\usage{
print_named_list(my_list)
}
\arguments{
\item{my_list}{A named list to be converted to a string.}
}
\value{
A string with each element of the list on a separate line and with its name.
}
\description{
This function converts a named list to a string, with each element of the list on a separate line and with its name.
}
\examples{
my_list <- list(a = 1, b = 2, c = 3)
print_named_list(my_list)
}
