% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_date_rng.R
\name{quick_date_rng}
\alias{quick_date_rng}
\title{Quick Date Range}
\usage{
quick_date_rng(mon = 6)
}
\arguments{
\item{mon}{integer input.}
}
\value{
vector of dates
}
\description{
Function returns min and max date values given month input by user.
Mostly, to create date ranges for plots.
}
\examples{
quick_date_rng()

quick_date_rng(month = 3)
}
