% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_leaflet.R
\name{quick_leaflet}
\alias{quick_leaflet}
\title{Make a quick leaflet plot.}
\usage{
quick_leaflet(data, markers = F, lines = F, polys = F)
}
\arguments{
\item{data}{a spatial object}

\item{markers}{boolean to indicate if markers should be used - default value is `false`}

\item{lines}{boolean to indicate if lines should be used - default value is `false`}

\item{polys}{boolean to indicate if polygons should be used - default value is `false`}
}
\value{
a leaflet map
}
\description{
Make a quick leaflet plot.
}
\examples{
mapview::breweries \%>\%
 quick_leaflet(markers = T)
}
