% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_import.R
\name{read_csv_allFiles2}
\alias{read_csv_allFiles2}
\title{Read multiple csv files in singular folder.}
\usage{
read_csv_allFiles2(
  data_location = "data",
  specifically = NULL,
  clean = F,
  clean_string = NULL,
  latest = F
)
}
\arguments{
\item{data_location}{a string that defines which folder data is in - omit trailing `/`. Default is set to `/data/`.}

\item{specifically}{a string used to reprex filter files - use to limit which files are read.}

\item{clean}{a boolean to clean the data - BROKEN, always keep as default `FALSE`.}

\item{clean_string}{a string indicating a string that should be removed from file name when naming list entries.}

\item{latest}{a boolean (T/F) indicating if the function should only load the latest data - usually the latest by file name string date.}
}
\value{
a named list of dataframes
}
\description{
This function can be used to read many files at once.
Intended to read the same file that has been written multiple times using different date suffix.
Example: same data that is queried daily and needs to be saved bu analyzed together.
}
\examples{
#none
}
