% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind_me_packages_to_use.R
\name{remind_me_packages_to_use}
\alias{remind_me_packages_to_use}
\title{Print basic packages that to use template}
\usage{
remind_me_packages_to_use()
}
\value{
None (prints to console).
}
\description{
Print basic packages that to use template
}
\examples{
remind_me_packages_to_use()

}
