% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindMe_gitIgnore.R
\name{remindMe_gitIgnore}
\alias{remindMe_gitIgnore}
\title{Print basic to console basic tryCatch template}
\usage{
remindMe_gitIgnore()
}
\value{
None (prints to console).
}
\description{
Print basic to console basic tryCatch template
}
\examples{
remindMe_gitIgnore()

}
