% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindMe_packages_to_use.R
\name{remindMe_packages_to_use}
\alias{remindMe_packages_to_use}
\title{Print basic packages that to use template}
\usage{
remindMe_packages_to_use()
}
\value{
None (prints to console).
}
\description{
Print basic packages that to use template
}
\examples{
remindMe_packages_to_use()

}
