% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindMe_trycatch.R
\name{remindMe_trycatch}
\alias{remindMe_trycatch}
\title{Print basic to console basic tryCatch template}
\usage{
remindMe_trycatch()
}
\value{
None (prints to console).
}
\description{
Print basic to console basic tryCatch template
}
\examples{
remindMe_trycatch()

}
