% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_prompt_used.R
\name{robust_prompt_used}
\alias{robust_prompt_used}
\title{Ask user for T/F input with user defined prompt.}
\usage{
robust_prompt_used(prompt)
}
\arguments{
\item{prompt}{A character string indicating the action to be performed.}
}
\value{
A logical value indicating whether or not the user elected to perform
the specified action. Returns TRUE if the user entered "Y", and FALSE if the
user entered "N".
}
\description{
This function prompts the user with a message asking whether or not they want
to perform a specific action. The function accepts the user's response and
handles invalid input by prompting the user again.
}
\examples{
\dontrun{
robust_prompt_used("continue with the analysis")
}
}
