% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_spatial_sf.R
\name{self_buffer}
\alias{self_buffer}
\title{self_buffer}
\usage{
self_buffer(object, rad_bg = NA, rad_sm = NA, nm = NULL)
}
\arguments{
\item{object}{an sf object (generally in lat/long format)}

\item{rad_bg}{a numeric value}

\item{rad_sm}{a numeric value}

\item{nm}{}
}
\value{
a list containing two sf buffer polygons
}
\description{
Creates two buffer polygons around a line or polygon that act together to make a negative space buffer - image a donut shape around a line. It can be used to help geo-reference/identify unknown links to a known network. It creates a poly buffer AROUND a line (roadway) and can be used to filter things EXCEPT the line
}
\examples{
#work in progress
}
