% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_shiny.R
\name{shiny_box_common}
\alias{shiny_box_common}
\title{Create a standardized box for UI elements in Shiny applications}
\usage{
shiny_box_common(title = NA, object_in_box = NA, collapsed = FALSE)
}
\arguments{
\item{title}{The title of the box (default is NA).}

\item{object_in_box}{The UI elements to be placed inside the box (default is NA).}

\item{collapsed}{Logical indicating whether the box is initially collapsed (default is FALSE).}
}
\value{
A standardized Shiny box with specified parameters.
}
\description{
This function generates a standard box for UI elements in Shiny applications.
It includes commonly used parameters, making it convenient to create consistent
box structures without repeatedly specifying the same inputs.
}
\examples{
\dontrun{
  # Create a standard box with a title and UI elements
  shiny_box_common(title = "Example Box", object_in_box = fluidRow(plotOutput("plot")))
}

}
\seealso{
\code{\link[shinydashboard]{box}}
}
