% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_shiny.R
\name{shiny_bs_popover}
\alias{shiny_bs_popover}
\title{Create a Shiny button with popover message}
\usage{
shiny_bs_popover(id, title, text, trigger = "hover")
}
\arguments{
\item{id}{The unique identifier for the button and popover.}

\item{title}{The title of the popover message.}

\item{text}{The text content of the popover message.}

\item{trigger}{The trigger for displaying the popover (default is "hover").}
}
\value{
A Shiny tagList containing the button and popover components.
}
\description{
This function generates a Shiny button along with a popover message. The
button, when pressed or hovered over (based on the specified trigger), displays
a popover message with the provided title and text. The default behavior is
to show an info-style message with a question mark icon.
}
\examples{
\dontrun{
  shiny_bs_popover(id = "example_button", title = "Information",
                   text = "This is an example popover message.")
}

}
\seealso{
\code{\link[shinyBS]{bsButton}}, \code{\link[shinyBS]{bsPopover}},
\code{\link[shiny]{icon}}
}
