% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_shiny.R
\name{shiny_get_list_items}
\alias{shiny_get_list_items}
\title{Extract specific elements from the Shiny input list}
\usage{
shiny_get_list_items(input_list, suffix = NA, string, purrr = TRUE)
}
\arguments{
\item{input_list}{The Shiny input list to extract elements from.}

\item{suffix}{The optional suffix to filter input elements.}

\item{string}{The string pattern to filter input elements.}

\item{purrr}{Logical indicating whether to use purrr-style functions for filtering.}
}
\value{
A subset of the input list containing the matching elements.
}
\description{
This function takes the master Shiny input list at any given time and
extracts the list elements by matching their names based on a suffix and/or
a string pattern. It provides a convenient way for users to obtain the
instantaneous values of specific input elements for troubleshooting purposes.
}
\examples{
\dontrun{
  # Extract input elements with a suffix and matching a string pattern
  shiny_get_list_items(input_list = input, suffix = "numeric", string = "slider")
}

}
\seealso{
\code{\link[stringr]{str_detect}}
}
