% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_shiny.R
\name{shiny_modal}
\alias{shiny_modal}
\title{Create a convenient Shiny modal for pop-up messages}
\usage{
shiny_modal(trigger, msg, size = "l")
}
\arguments{
\item{trigger}{The triggering event that activates the modal.}

\item{msg}{The message content to be displayed in the modal.}

\item{size}{The size of the modal ("s" for small, "m" for medium, "l" for large).}
}
\description{
This function wraps commonly used functions to create a Shiny modal for
pop-up messages. It allows you to link a triggering event with a message,
making it easier to display pop-up messages in Shiny applications with
reduced code repetition.
}
\examples{
\dontrun{
  # Create a Shiny modal linked to a triggering event with a message
  shiny_modal(trigger = input$button_click, msg = "Process completed successfully!")
}

}
\seealso{
\code{\link[shiny]{observeEvent}}, \code{\link[shiny]{showModal}},
\code{\link[shiny]{modalDialog}}
}
