% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_shiny.R
\name{shiny_spacer_row}
\alias{shiny_spacer_row}
\title{Create a blank spacer row for spacing between UI elements}
\usage{
shiny_spacer_row(size)
}
\arguments{
\item{size}{The height of the blank spacer row.}
}
\value{
A blank spacer row with the specified height.
}
\description{
This function generates a blank spacer row to create space between UI elements
in Shiny applications. It is a convenience function that inserts a fluid row
with an empty box having a specified height to create vertical space.
}
\examples{
\dontrun{
  # Create a spacer row with a specified height
  shiny_spacer_row(size = "20px")
}

}
\seealso{
\code{\link[shiny]{fluidRow}}, \code{\link[shinydashboard]{box}}
}
