% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_clean_date.R
\name{strg_clean_date}
\alias{strg_clean_date}
\title{Make a clean date string of current date with simple punctuation and syntax.}
\usage{
strg_clean_date()
}
\value{
character string of date without bad punctuation
}
\description{
Make a clean date string of current date with simple punctuation and syntax.
}
\examples{

clean_date()
}
