% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_strings.R
\name{strg_clean_datetime}
\alias{strg_clean_datetime}
\title{Make a clean datetime string of current datetime with simple punctuation and syntax.}
\usage{
strg_clean_datetime()
}
\value{
character string of datetime using only underscores
}
\description{
Make a clean datetime string of current datetime with simple punctuation and syntax.
}
\examples{

clean_datetime()
}
