% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_strings.R
\name{strg_col_clean}
\alias{strg_col_clean}
\title{Quickly clean a dataframe's column names}
\usage{
strg_col_clean(df, na_marker = "")
}
\arguments{
\item{df}{A data frame to clean.}

\item{na_marker}{A value that should be replaced with NA.}
}
\value{
The cleaned data frame.
}
\description{
This function cleans a data frame by removing empty rows and columns, converting column names to snake_case, and replacing a specified value with NA.
}
\examples{
df <- data.frame(A = c(1, NA, 3), B = c("apple", "orange", ""), C = c("", "banana", ""))
strg_col_clean(df, "")

}
