% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_extract_delim_pos.R
\name{strg_extract_delim_pos}
\alias{strg_extract_delim_pos}
\title{Extract a position from a delimiter-separated string}
\usage{
strg_extract_delim_pos(string, pos, delim = "/")
}
\arguments{
\item{string}{A character string to split.}

\item{pos}{Integer position to extract, counted from the end of the string.
For example, \code{pos = 1} returns the last element, \code{pos = 2} the
second-to-last, etc.}

\item{delim}{Delimiter used to split the string. Defaults to \code{"/"}.}
}
\value{
A character string corresponding to the requested position.
}
\description{
Splits a string by a specified delimiter and returns the element at a given
position counted from the **end** of the string. This is useful for reliably
extracting filenames, directory names, or other path components from
consistently delimited strings.
}
\details{
This function assumes the input string is regularly delimited and does not
perform bounds checking. If \code{pos} exceeds the number of available
elements, \code{NA} will be returned.
}
\examples{
\dontRun{
path <- "C:/Users/GauntM/Documents/010_projects/replicaToolkitR_lab2/projects/replica_md_quick_builds/data/data_thru_bbox_20260106_200951068395_2024_Q2_test"

# Get the last path component
strg_extract_delim_pos(path, pos = 1, delim = "/")

# Get the parent directory name
strg_extract_delim_pos(path, pos = 2, delim = "/")
}

}
