% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_fun_eval.R
\name{strg_fun_eval}
\alias{strg_fun_eval}
\title{Evaluate a string as a command}
\usage{
strg_fun_eval(text)
}
\arguments{
\item{text}{A character string representing the R command to evaluate.}
}
\value{
The result of evaluating the R command.
}
\description{
Takes a string and evaluates it as an R command.
}
\examples{
\dontrun{
strg_fun_eval("1 + 1")
strg_fun_eval("mean(c(1, 2, 3))")

}
}
