% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_numeric_order.R
\name{strg_numeric_order}
\alias{strg_numeric_order}
\title{Order factor levels numerically}
\usage{
strg_numeric_order(col, rev = TRUE)
}
\arguments{
\item{col}{A factor or character vector.}

\item{rev}{Logical; if `TRUE`, reverse the order of levels.}
}
\value{
A factor with reordered levels.
}
\description{
`strg_numeric_order` orders factor levels numerically based on the values in the input vector.
If `rev` is set to `TRUE`, the levels are reversed.
}
\details{
This function is useful for ensuring that factor levels are ordered correctly when dealing with numeric values.
It leverages the `stringr::str_sort` function to sort the levels.
}
\examples{
\dontrun{
temp_df = paste0("lab_", seq(1, 50, 2)) \%>\%
data.frame(label_unordered = .
           ,label_ordered = strg_numeric_order(.))

temp_df \%>\% arrange(label_unordered) \%>\% print()
temp_df \%>\% arrange(desc(label_ordered)) \%>\% print()

}
}
