% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_pretty_char.R
\name{strg_pretty_char}
\alias{strg_pretty_char}
\title{Make a character string pretty}
\usage{
strg_pretty_char(col)
}
\arguments{
\item{col}{A character string to make pretty.}
}
\value{
The pretty character string.
}
\description{
This function replaces underscores with spaces and capitalizes the first letter of each word in a character string.
}
\details{
Mostly used when making colnames or items in table pretty.
}
\examples{
\dontrun{
strg_pretty_char("hello_world")
strg_pretty_char("the_quick_brown_fox")

}
}
