% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_pretty_num.R
\name{strg_pretty_num}
\alias{strg_pretty_num}
\title{Make a number pretty for printing}
\usage{
strg_pretty_num(vector, rnd = 0)
}
\arguments{
\item{vector}{A numeric vector to make pretty.}

\item{rnd}{The number of digits to round to.}
}
\value{
The pretty number string.
}
\description{
This function formats a number to make it more readable. Large numbers are expressed in millions or billions and small numbers are not rounded.
}
\examples{
\dontrun{
strg_pretty_num(1234)
strg_pretty_num(12345678)
strg_pretty_num(-987654321, 2)

}
}
