% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_targets.R
\name{tar_scan_file}
\alias{tar_scan_file}
\title{Scan File for Target String}
\usage{
tar_scan_file(file, print_tar = FALSE, rm_tar_load = T)
}
\arguments{
\item{file}{A character string specifying the path to the file to be scanned.}

\item{print_tar}{Logical; indicating whether to print the lines in a format suitable for tar loading. Default is \code{FALSE}.}

\item{rm_tar_load}{Logical; indicating whether to remove remove lines with tar_load{tar_objects}. Users may load objects but they are not used in the report, removing these items will give a list of tar objects that are actually used in the analysis. Default is \code{TRUE}.}
}
\value{
A character vector containing lines from the file matching the criteria specified.
}
\description{
This function scans a file for lines containing a specific target string and performs optional manipulations on the lines.
}
\examples{
\dontrun{
# Scan a file for lines containing "plt_mdl_" and "plt_agg" strings, remove lines containing "tar_read",
# remove trailing spaces, and print lines in tar loading format.
tar_scan_file("example.txt", "plt_mdl_|plt_agg"), rm_tar = TRUE, rm_spce = TRUE, print_tar = TRUE)
}

}
