% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_mad.R
\name{ts_mad}
\alias{ts_mad}
\title{Make rolling MAD (median absolute deviation) values for dataframe attribute.}
\usage{
ts_mad(data, window = 20)
}
\arguments{
\item{data}{a data frame with a timerseries attribute - attribute must be named value to work}

\item{window}{an integer indicating width of window}
}
\value{
a dataframe with rolling MAD statistic attribute
}
\description{
Make rolling MAD (median absolute deviation) values for dataframe attribute.
}
\examples{

\dontrun{
test_data = data.frame(index = 1:53,
                      value = c(rnorm(25, 10, 3)
                                ,rnorm(3, 100, 3)
                                ,rnorm(25, 10, 3)))
test_data \%>\%
 ts_mad(window = 7)
 }
}
