% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_zscore.R
\name{ts_zscore}
\alias{ts_zscore}
\title{Make rolling zscore attribute from timeseries.}
\usage{
ts_zscore(data, grp = NULL, width = 7, align = "right", threshold = 3)
}
\arguments{
\item{data}{a dataframe containing the timeseries attribute}

\item{grp}{character list of attributes to group by}

\item{width}{integer value indicating widtch of window}

\item{align}{character indicating which side of value window will be aligned. Default is `right`, which only uses past values. Other options are `center` and `left`.}

\item{threshold}{integer indicating zscore away from center function will flag value as an outlier}
}
\value{
a dataframe with zcore and zscore_flag attributes
}
\description{
Make rolling zscore attribute from timeseries.
}
\examples{
data.frame(index = 1:60, value = c(rnorm(50, 10, 3), rnorm(10, 25, 0))) \%>\%
 na.omit() \%>\%
 ts_zscore(width = 30)

c(10, 30) \%>\%
 map(~data.frame(index = 1:60
                 ,value = c(rnorm(50, 10, 3), rnorm(10, 25, 0))) \%>\%
       na.omit() \%>\%
       ts_zscore(width = .x)) \%>\%
 reduce(bind_rows)
}
