% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_manip.R
\name{dgt0}
\alias{dgt0}
\title{Round to 0 digits}
\usage{
dgt0(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Use this function to round a vector.
}
\examples{
dgt0(rnorm(10))
}
