% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor_divide.R
\name{floor_divide}
\alias{floor_divide}
\title{Floor divide a vector by some value.}
\usage{
floor_divide(value, floor)
}
\arguments{
\item{value}{A numeric value or dataframe column to be floor divided}

\item{floor}{A value to floor divide by}
}
\value{
A numeric value or vector
}
\description{
This function can be applied to a vector to bin values by a floor divide value.
}
\examples{

temp_index = rnorm(10, 100, 100)
floor_divide(temp_index, 5)
floor_divide(temp_index, 50)
floor_divide(temp_index, 100)
}
