% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_import.R
\name{gdrive_send_rds_files}
\alias{gdrive_send_rds_files}
\title{Send RDS Files to Google Drive}
\usage{
gdrive_send_rds_files(folder_with_files, gdrive_folder)
}
\arguments{
\item{folder_with_files}{The local directory path containing the RDS files to be sent. Other file types can be in this location - they will be filtered out.}

\item{gdrive_folder}{The folder on the GDrive you want data saved to - needs to be top level, do not make it multi-level.}
}
\value{
None (invisibly returns NULL)
}
\description{
This function programmatically sends RDS files saved in a user-provided directory
to a specified Google Drive location. The RDS files are converted to CSV format
before being uploaded to Google Drive.
}
\examples{
\dontrun{
# Provide the directory path containing the RDS files
folder_with_files <- "path/to/rds/files"

# Send the RDS files to Google Drive
gdrive_send_rds_files(folder_with_files)
}

}
\keyword{Drive,}
\keyword{Google}
\keyword{RDS,}
\keyword{file}
\keyword{upload}
