% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_spatial_sf.R
\name{quick_buffer}
\alias{quick_buffer}
\title{quick_buffer}
\usage{
quick_buffer(geo_data, to = 4326, with = 2781, radius = NA)
}
\arguments{
\item{geo_data}{an sf object (generally in lat/long format)}

\item{to}{a numeric CRS}

\item{with}{a  numeric CRS which allows for buffering}

\item{radius}{a numeric value of the buffer radius}
}
\value{
an sf mulitpolygon object
}
\description{
Easily buffer a spatial object. Primarily used to help convert between lat/long that cannot be buffered to CRS that can be and then back to lat/long.
}
\examples{
#work in progress
}
