% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strg_clean_datetime.R
\name{strg_clean_datetime}
\alias{strg_clean_datetime}
\title{Make a clean datetime string of current datetime with simple punctuation and syntax.}
\usage{
strg_clean_datetime(strip = FALSE)
}
\arguments{
\item{strip}{A boolean to indicate where or not to truncate the datetime string to 15 characters, default F.}
}
\value{
character string of datetime using only underscores
}
\description{
Make a clean datetime string of current datetime with simple punctuation and syntax.
}
\examples{
\dontrun{

clean_datetime()
}
}
